week_7_ecr/requirements_inference.txt
pytorch-lightning==1.2.10
datasets
scikit-learn
hydra-core
omegaconf
hydra_colorlog
onnxruntime
fastapi
uvicorn
dvc
wandb